rename `instruction::ctxt_ctxt_equality` to non-abbreviated name
