+++ zk-token-sdk/src/instruction/ciphertext_commitment_equality.rs
@@ -13,7 +13,7 @@ use {
             pedersen::{PedersenCommitment, PedersenOpening},
         },
         errors::ProofError,
-        sigma_proofs::ctxt_comm_equality_proof::CiphertextCommitmentEqualityProof,
+        sigma_proofs::ciphertext_commitment_equality_proof::CiphertextCommitmentEqualityProof,
         transcript::TranscriptProtocol,
     },
     merlin::Transcript,
