+++ zk-token-sdk/src/instruction/mod.rs
@@ -1,6 +1,6 @@
 pub mod batched_range_proof;
+pub mod ciphertext_ciphertext_equality;
 pub mod ciphertext_commitment_equality;
-pub mod ctxt_ctxt_equality;
 pub mod pubkey_validity;
 pub mod range_proof;
 pub mod transfer;
@@ -27,11 +27,11 @@ pub use {
         batched_range_proof_u64::BatchedRangeProofU64Data, BatchedRangeProofContext,
     },
     bytemuck::Pod,
+    ciphertext_ciphertext_equality::{
+        CiphertextCiphertextEqualityProofContext, CiphertextCiphertextEqualityProofData,
+    },
     ciphertext_commitment_equality::{
         CiphertextCommitmentEqualityProofContext, CiphertextCommitmentEqualityProofData,
-    },
-    ctxt_ctxt_equality::{
-        CiphertextCiphertextEqualityProofContext, CiphertextCiphertextEqualityProofData,
     },
     pubkey_validity::{PubkeyValidityData, PubkeyValidityProofContext},
     range_proof::{RangeProofContext, RangeProofU64Data},
